+++ src/ralph_assets/models_support.py
@@ -6,6 +6,8 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
+from datetime import datetime
+from django.contrib.humanize.templatetags.humanize import naturaltime
 from django.core.urlresolvers import reverse
 from django.db import models
 from django.utils.translation import ugettext_lazy as _
@@ -17,6 +19,7 @@ from lck.django.common.models import (
     TimeTrackable,
     WithConcurrentGetOrCreate,
 )
+
 from ralph.discovery.models_util import SavingUser
 from ralph_assets import models_assets
 from ralph_assets.models_assets import (
@@ -88,6 +91,9 @@ class Support(
     support_type = models.ForeignKey(
         SupportType,
         on_delete=models.PROTECT,
+        blank=True,
+        default=None,
+        null=True,
     )
     assets = models.ManyToManyField(Asset, related_name='supports')
 
@@ -101,3 +107,6 @@ class Support(
             'support_id': self.id,
             'mode': ASSET_TYPE2MODE[self.asset_type],
         })
+
+    def get_natural_end_support(self):
+        return naturaltime(datetime(*(self.date_to.timetuple()[:6])))